simplify converting point lists to shapes
I ran into a lot of bugs with the previous implementation, maybe on
around 5-10% of files. This one seems to work on almost all files.
There's still one bug with MLP422_593, where a path hits a dead end
before seeing a cycle. It seems to affect less than 1% of files though.